restructure command set
