+++ cmds_n_settings.md
@@ -7,15 +7,11 @@
 ## Ch-ch-ch-changes
 
 - [X] add c[lear] 
+- [X] re-arrange commands in help menu
+  - before: in alphabetical order 
+  - now: in order of registration
+
 - write top-level user manual
-- re-arrange commands in help menu
-  - now in alphabetical order 
-  - then in chosen order [e.g. order of registration]
-
-:vis -pdf -v -token -eval -env -paste
-:vis -cons -vv
-
-
 
 
 ## new Commands:
@@ -26,41 +22,40 @@
 +----------------+-------------------+----------------------------------------------------------+
 
 GENERAL:
-| h[elp]         | ~                 | list all commands with usage                             |
+| h[elp]         | ~                 | list all commands with usage                             |1
 |                | ~ <cmd>           | print usage command <cmd>                                |
-| q[uit]         | ~                 | quit the session                                         |
-| cl[earscreen]  | ~                 | clear the terminal screen                                |
+| q[uit]         | ~                 | quit the session                                         |2
+| cl[earscreen]  | ~                 | clear the terminal screen                                |3
 
 ENVIRONMENT:
-| c[lear]        | ~                 | clear the environment                                    | # c added
-| l[ist]         | ~                 | list all identifiers in the environment alphabetically   | -- dependent on verbosity; maybe also on inclenvironment
+| l[ist]         | ~                 | list all identifiers in the environment alphabetically   | 4 -- dependent on verbosity; maybe also on inclenvironment
 |                |                   |      with types and values                               |
-
-SETTINGS:
-| settings       | ~                 | list all settings with their current values and defaults |
-|                |                   |      for an overview consult :settings and/or :h set     |
-| set            | ~ ...             |                                                          |
-|
-| reset          | ~ <setting>       | set <setting> to default                                 |
-| unset          | ~ <setting>       | set boolean <setting> to false                           |
-|                |                   |      for an overview consult :settings and/or :h set     |
+| c[lear]        | ~                 | clear the environment                                    | 5 # c added
 
 PASTE:
-| paste          | ~ <input>         | evaluate multiline <input> (terminated by blank line)    |
+| paste          | ~ <input>         | evaluate multiline <input> (terminated by blank line)    | 6
 
 LEVEL:
-| expr[ession]   | ~ <input>         | expect <input> to be an expression                       |
-| stmt|statement | ~ <input>         | expect <input> to be a statement                         |
-| prog[ram]      | ~ <input>         | expect <input> to be a program                           |
+| expr[ession]   | ~ <input>         | expect <input> to be an expression                       | 7
+| stmt|statement | ~ <input>         | expect <input> to be a statement                         | 8 
+| prog[ram]      | ~ <input>         | expect <input> to be a program                           | 9
 
 
 PROCESS:
-| e[val]         | ~ <input>         | print out value of object <input> evaluates to           |
-| p[arse]        | ~ <input>         | parse <input>                                            |
-| t[ype]         | ~ <input>         | show objecttype <input> evaluates to                     |
-| trace          | ~ <input>         | show evaluation trace step by step                       |
-| parsetree      | ~ <input>         | show parsetree                                           |
-| evaltree       | ~ <input>         | show evaltree                                            |
+| p[arse]        | ~ <input>         | parse <input>                                            | 10
+| p[arse]tree    | ~ <input>         | show parsetree                                           | 11
+| e[val]         | ~ <input>         | print out value of object <input> evaluates to           | 12
+| t[ype]         | ~ <input>         | show objecttype <input> evaluates to                     | 13
+| trace          | ~ <input>         | show evaluation trace step by step                       | 14
+| e[val]tree     | ~ <input>         | show evaltree                                            | 15
+
+SETTINGS:
+| settings       | ~                 | list all settings with their current values and defaults | 16
+|                |                   |      for an overview consult :settings and/or :h set     |
+| set            | ~ ...             | ...                                                      | 17
+| reset          | ~ <setting>       | set <setting> to default                                 | 18
+| unset          | ~ <setting>       | set boolean <setting> to false                           | 19
+|                |                   |      for an overview consult :settings and/or :h set     |
 
 ```
